requirements.txt
certifi==2019.9.11
chardet==3.0.4
cloudinary==1.18.2
dataclasses==0.6
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-ckeditor==5.8.0
django-cors-headers==3.2.0
django-js-asset==1.2.2
djangorestframework==3.10.3
djongo==1.2.36
dnspython==1.16.0
environs==6.1.0
gunicorn==19.9.0
idna==2.8
marshmallow==3.2.2
mock==3.0.5
mongoengine==0.18.2
oauthlib==3.1.0
Pillow==6.2.1
pymongo==3.9.0
python-decouple==3.3
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
sqlparse==0.2.4
urllib3==1.25.7
whitenoise==4.1.2
